requirements.txt
# Нужно для работы GRPC сервера
grpcio==1.14.1
gcloud==0.18.3
google-api-python-client==1.7.3
grpcio_health_checking==1.16.0
